Fixed bgm and added exit
