+++ app/src/main/AndroidManifest.xml
@@ -8,11 +8,22 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
-        <activity android:name=".views.createRegionActivity"/>
-        <activity android:name=".views.displayPlayerActivity" />
-        <activity android:name=".views.selectDifficultyActivity" />
-        <activity android:name=".views.createPlayerActivity" />
+        android:theme="@style/AppTheme.NoActionBar">
+        <activity
+            android:name=".views.createRegionActivity">
+        </activity>
+        <activity
+            android:name=".views.displayPlayerActivity"
+            android:theme="@style/AppTheme">
+        </activity>
+        <activity
+            android:name=".views.selectDifficultyActivity"
+            android:theme="@style/AppTheme">
+        </activity>
+        <activity
+            android:name=".views.createPlayerActivity"
+            android:theme="@style/AppTheme">
+        </activity>
         <activity
             android:name=".views.MainActivity"
             android:label="@string/app_name"
