+++ app/src/main/res/values/strings.xml
@@ -23,5 +23,6 @@
     <string name="confirm">Right...\nSo your name is</string>
     <string name="yes">Confirm</string>
     <string name="no">Cancel</string>
+    <string name="exit">Exit</string>
 </resources>
